Enable LUKS service to start at boot; make ZFS service require it
I wasn't able to enable the service until I put in an [Install] section,
which I guess makes sense!

Another option would have been to leave out that section and not bother
enabling it, since the commdata-zfs service relies on it and *that*
one has an [Install] section and will be enabled for startup at boot.