Just use roles to assign services -- give up on groups
Groups provided an acceptable level of indirection, but when I changed
multiple roles to add an explicit dependency on commdata (for safety),
the playbook run would run the commdata role multiple times on the same
host. This seems to be because the host was in multiple groups, and role
runs were not deduplicated across group runs.

- Rolled dyndns secret and moved dyndns role to murphy, which is the
  future supervisor.
- Moved service secrets from group_vars to role vars. Vault loading is now
  explicit since the vars.yml/vault.yml pattern doesn't work in role vars.

Also tweaked sandstorm's dependencies to include terminate-tls.