Fix non-converging enabling of LUKS service
Possibly related to change in 5edf4d44d to make commdata-fs `Requires`
commdata-luks, every time I ran Ansible on murphy it was saying that the
task to enable the LUKS service had "changed".

When I disabled the service, I wasn't able to reenable it! It seems like
systemd was deleting the symlink from /etc/systemd/system to the file in
/opt/commapps.

This change moves the file to be directly in /etc/systemd/system, and it
appears to have the desired effect -- the build converges, and the service
can be disabled and re-enabled successfully.